rename: 스토리북 문서 위치 이동
